apps/python/tests/test_placeholder.py
"""Placeholder tests for react_agent_with_mcp_tools module.

These tests will be expanded as the react_agent_with_mcp_tools module develops.
For now, they ensure the CI pipeline has something to run.
"""

import pytest


class TestToolsAgentPlaceholder:
    """Placeholder test class for react_agent_with_mcp_tools."""

    def test_import_react_agent_with_mcp_tools(self) -> None:
        """Test that react_agent_with_mcp_tools module can be imported."""
        from react_agent_with_mcp_tools import agent

        assert agent is not None

    def test_graph_exists(self) -> None:
        """Test that the graph object exists in agent module."""
        from react_agent_with_mcp_tools.agent import graph

        assert graph is not None

    @pytest.mark.skip(reason="Requires LLM configuration")
    def test_graph_invocation(self) -> None:
        """Test that graph can be invoked (requires LLM setup)."""
        # This test is skipped by default as it requires LLM configuration
        # It serves as a template for future integration tests
        pass
