apps/python/pyproject.toml
[project]
name = "fractal-agents-runtime"
version = "0.0.0"
description = "Free, self-hostable LangGraph agent runtime with MCP tools, RAG, and A2A protocol"
authors = [
    { name = "l4b4r4b4b4" },
]
requires-python = ">=3.11.0,<3.13"
dependencies = [
    "langgraph>=1.0.8",
    "langchain-anthropic>=1.3.3",
    "langchain-core>=1.2.11",
    "langchain-openai>=1.1.9",
    "pydantic==2.11.3",
    "langchain>=1.2.10",
    "mcp>=1.9.1",
    "supabase>=2.15.1",
    "aiohttp>=3.8.0",
    "deepagents>=0.3.9",
    "tavily-python>=0.7.20",
    "robyn>=0.76.0",
    "apscheduler>=3.11.2",
    "croniter>=6.0.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary,pool]>=3.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langfuse>=3.14.1",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["react_agent_with_mcp_tools*", "robyn_server*"]
exclude = ["robyn_server.tests*", "robyn_server.helm*"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["src/robyn_server/tests"]
norecursedirs = ["archive", ".agent", ".venv", "__pycache__"]

[tool.ruff]
src = ["src"]
exclude = [
    ".agent",
    ".ruff_cache",
    ".venv",
    "__pycache__",
]
